Mesh, Storage: Moving these two into pde_solver class.
This is the first commit in a series of refactoring in order to bundle all quantities inside the pde_solver class.
The goal is to have a solid platform for use within simulators such as ViennaSHE and ViennaMini.
As these simulators may be GUI-driven, static dispatches as well as unnecessary book-keeping needs to be reduced to a minimum.